reddit/requirements.txt
langchain 
langchain_openai 
praw 
openai 
python-dotenv
tweepy
motor
tweepy